+++ R/search_games.R
@@ -13,6 +13,9 @@
 #' @param query_file An optional argument specifying the file storing query of
 #'   games. If leave as `NULL`, default to "sql/games.tmpl.sql", which is
 #'   created by rmarkdown template.
+#' @param file_cache The file used to store caches of results which is passed to
+#'   [cachem::cache_disk()].
+#' @param ... Arguments passed to [search_games()].
 #' @return A [tibble][tibble::tibble-package] contains all the games to be
 #'   analyzed and its related information.
 #' @export
@@ -20,10 +23,19 @@ search_games <- function(config_where, known_only = TRUE, query_file = NULL) {
   if (is.null(query_file)) query_file <- "sql/games.tmpl.sql"
   if (!file.exists(query_file)) abort("Query file missing.", "query_file_miss")
   games <- pickup(query_file, config_where)
+  if (bit64::is.integer64(games$game_id)) {
+    games$game_id <- bit64::as.character.integer64(games$game_id)
+  }
   if (known_only) {
     games |>
       dplyr::inner_join(data.iquizoo::game_info, by = "game_id") |>
-      dplyr::mutate(prep_fun = syms(.data[["prep_fun_name"]]))
+      dplyr::mutate(
+        prep_fun = syms(.data[["prep_fun_name"]]),
+        dplyr::across(
+          dplyr::all_of(c("input", "extra")),
+          parse_exprs
+        )
+      )
   } else {
     games |>
       dplyr::left_join(data.iquizoo::game_info, by = "game_id")
@@ -33,7 +45,9 @@ search_games <- function(config_where, known_only = TRUE, query_file = NULL) {
 #' @describeIn search_games Cached version using
 #'   [memoise()][memoise::memoise()].
 #' @export
-search_games_mem <- memoise::memoise(
-  search_games,
-  cache = cachem::cache_disk("~/.cache.tarflow")
-)
+search_games_mem <- function(file_cache = "~/.cache.tarflow", ...) {
+  memoise::memoise(
+    search_games,
+    cache = cachem::cache_disk(file_cache)
+  )(...)
+}