+++ tests/testthat/test-search_games.R
@@ -1,5 +1,4 @@
 test_that("Correctly combine `game_info`", {
-  skip_on_os("windows")
   mockery::stub(
     search_games, "pickup",
     function(query_file, ...) tibble::tibble(read.csv(query_file))
@@ -10,6 +9,18 @@ test_that("Correctly combine `game_info`", {
   )
   expect_snapshot_value(
     search_games(known_only = FALSE, query_file = "dummy/test.sql"),
+    style = "json2"
+  )
+})
+
+test_that("Support `integer64`", {
+  mockery::stub(
+    search_games, "pickup",
+    function(query_file, ...)
+      tibble::tibble(game_id = bit64::as.integer64.double(225528186135045))
+  )
+  expect_snapshot_value(
+    search_games(query_file = "dummy/test.sql"),
     style = "json2"
   )
 })